Use indefinite wait instead of wait_for
Previously, the producer and consumer threads used wait_for. This causes
the thead to peridically wake up, grab the lock and check the predicate.
But, in the program, we can use the indefinite wait without issues as the
watcher thread will notify the threads that a shutdown is requested.

Use of the wait_for is sometimes needed when we cannot be sure that the
thread will be signalled eventually. This can happen due to a bug in a
complex software or if for some reason the notifier thread dies due to
some issue without seding the notify signal.